+++ Application/machine_learning/classification/weighted_logistic_regression_l2_norm.py
@@ -35,7 +35,7 @@ class WeightedLogisticRegressionL2Norm:
             label (numpy.array): The label of a dataset.
             initial_coefficients (numpy.array): Initial weights for the model.
             weights_list (numpy.array): List of weights
-            step_size (int): Step size.
+            step_size (float): Step size.
             max_iter (int): Amount of iterations.
             l2_penalty (float): L2 penalty value.
